Added some more recipes. Increased robustness of database shell script

+++ db/gen_db.sh
@@ -1,7 +1,21 @@
 #!/bin/bash
 
 cd `dirname $0`
-rm caffeine_db
+
+if [ -e caffeine_db ];
+then
+	echo "caffeine_db exists. Removing"
+	rm caffeine_db
+fi
+
+echo "Creating caffeine_db..."
 sqlite3 caffeine_db < importcmd.txt
-mkdir ../assets/
+
+if [ ! -d ../assets ];
+then
+	echo "Assets folder doesn't exist. Creating"
+	mkdir ../assets
+fi
+
+echo "Copying caffeine_db to assets folder..."
 cp caffeine_db ../assets/
